server/requirements.txt
amqp==2.2.2
astroid==1.5.3
autopep8==1.3.2
awsebcli==3.12.1
billiard==3.5.0.3
blessed==1.14.2
botocore==1.8.20
cement==2.8.2
colorama==0.3.7
Django==2.0
django-common-helpers==0.9.1
django-environ==0.4.3
django-extensions==1.9.8
django-rest-swagger==2.1.2
djangorestframework==3.7.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
isort==4.2.15
jmespath==0.9.3
kombu==4.1.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.13.1
openapi-codec==1.3.2
pathspec==0.5.0
pip-autoremove==0.9.0
pkg-resources==0.0.0
psycopg2==2.7.3
pycodestyle==2.3.1
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.9.1
scikit-learn==0.19.0
scipy==0.19.1
semantic-version==2.5.0
simplejson==3.13.2
six==1.10.0
tabulate==0.7.5
termcolor==1.1.0
typing==3.6.2
vine==1.1.4
wcwidth==0.1.7
websocket-client==0.45.0
wrapt==1.10.11
